Change challenge selector to just be used for adding custom challenges
